requirements.txt
# ── Core ML ───────────────────────────────────────────────────────────────────
sentence-transformers>=2.7.0       # SBERT embeddings
transformers>=4.40.0               # BERT NLI (Layer 3) + AI detection
torch>=2.0.0                       # PyTorch backend (CPU is fine for hackathon)

# ── Vector search ─────────────────────────────────────────────────────────────
chromadb>=0.5.0                    # vector DB for corpus storage
faiss-cpu>=1.8.0                   # fast nearest-neighbor search (optional upgrade)

# ── Text extraction ───────────────────────────────────────────────────────────
PyMuPDF>=1.24.0                    # PDF extraction (import as fitz)
python-docx>=1.1.0                 # DOCX extraction

# ── NLP utils ─────────────────────────────────────────────────────────────────
nltk>=3.8.0                        # sentence tokenization
langdetect>=1.0.9                  # language detection
scikit-learn>=1.4.0                # TF-IDF (Layer 1), cosine similarity

# ── Backend ───────────────────────────────────────────────────────────────────
fastapi>=0.111.0                   # REST API
uvicorn[standard]>=0.29.0          # ASGI server
python-multipart>=0.0.9            # file upload support
python-dotenv>=1.0.1               # local .env support

# ── Database ──────────────────────────────────────────────────────────────────
pymongo>=4.7.0                     # MongoDB (embedding cache + submissions)

# ── Report generation ─────────────────────────────────────────────────────────
reportlab>=4.1.0                   # PDF report export

# ── HTTP requests (corpus fetching) ──────────────────────────────────────────
requests>=2.31.0
